Social footer (#1170)
* Use component library components in the social footer

* reset translations

* reset translations
